+++ src/main/java/com/example/budongbudong/domain/auction/dto/response/AuctionInfoResponse.java
@@ -4,21 +4,22 @@ import com.example.budongbudong.common.entity.Auction;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
 @RequiredArgsConstructor
 public class AuctionInfoResponse {
-    private final Long startPrice;
-    private final Long minBidIncrement;
-    private final Long highestPrice;
+    private final BigDecimal startPrice;
+    private final BigDecimal minBidIncrement;
+    private final BigDecimal highestPrice;
     private final int totalBidders;
     private final LocalDateTime startedAt;
     private final LocalDateTime endedAt;
 
     public static AuctionInfoResponse from(
             Auction auction,
-            Long highestPrice,
+            BigDecimal highestPrice,
             int totalBidders
     ) {
         return new AuctionInfoResponse(
